% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-astr.R
\name{as_astr}
\alias{as_astr}
\title{Convert character vector to astr class}
\usage{
as_astr(x, n, ucase = FALSE)
}
\arguments{
\item{x}{A character vector which convert to a bstr object.}

\item{n}{A character vector which is name of x.}

\item{ucase}{A logical. If TRUE the x is converted to upper case. (default: FALSE)}
}
\description{
Convert character vector to astr class
}
\examples{
as_astr("Wqrld", "HELLQ", ucase = TRUE)

}
