% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-bstr.R
\name{as_bstr}
\alias{as_bstr}
\title{Convert character vector to biostring class}
\usage{
as_bstr(x, n)
}
\arguments{
\item{x}{Character vector}

\item{n}{Character vector}
}
\description{
Convert character vector to biostring class
}
