% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dstr.R
\name{as_dstr}
\alias{as_dstr}
\title{Convert character vector to dstr class}
\usage{
as_dstr(x, n)
}
\arguments{
\item{x}{Character vector}

\item{n}{Character vector}
}
\description{
Convert character vector to dstr class
}
