% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-astr.R
\name{astr}
\alias{astr}
\title{Constructing astr class object}
\usage{
astr(x, n, ucase = F)
}
\arguments{
\item{x}{A character vector which convert to a bstr object.}

\item{n}{A character vector which is name of x.}

\item{ucase}{A logical. If TRUE the x is converted to upper case. (default: FALSE)}
}
\description{
Constructing astr class object
}
\examples{
astr("Wqrld", "HELLQ", ucase = TRUE)

}
