% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_seq.R
\name{astr_rand_seq}
\alias{astr_rand_seq}
\title{Generate random astr sequences of desired lengths.}
\usage{
astr_rand_seq(n, length, pattern = "[ARNDCQEGHILKMFPSTWYVX]", seed)
}
\arguments{
\item{n}{single integer, nubmer of sequences}

\item{length}{integer vector, desired sequence lengths}

\item{pattern}{character vector specifying character classes}

\item{seed}{a single integer. specify seeds.}
}
\description{
Generate random astr sequences of desired lengths.
}
\examples{
astr_rand_seq(3, 20, seed = 1)
astr_rand_seq(3, 5:7, c("[arnd]", "[ARND]", "[X.]"), seed = 1)

}
