% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{Bio2bstr}
\alias{Bio2bstr}
\title{Convert sequence to Biostrings::*StringSet object}
\usage{
Bio2bstr(x)
}
\arguments{
\item{x}{sequence}
}
\description{
Convert sequence to Biostrings::*StringSet object
}
