% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-bstr.R
\name{bstr}
\alias{bstr}
\title{Constructing bstr class object}
\usage{
bstr(x, n, ucase = T)
}
\arguments{
\item{x}{Character vector}

\item{n}{Character vector}

\item{ucase}{Character vector}
}
\description{
Constructing bstr class object
}
