% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{bstr2BioString}
\alias{bstr2BioString}
\title{Convert sequence to Biostrings::*StringSet object}
\usage{
bstr2BioString(x)
}
\arguments{
\item{x}{sequence}
}
\description{
Convert sequence to Biostrings::*StringSet object
}
