% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr2df.R
\name{bstr2df}
\alias{bstr2df}
\alias{bstr2longdf}
\alias{df2bstr}
\alias{longdf2bstr}
\title{bstrobj <-> data.frame}
\usage{
bstr2df(bstrobj)

bstr2longdf(bstrobj)

df2bstr(df)

longdf2bstr(longdf)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}

\item{df}{data.frame}

\item{longdf}{long data.frame}
}
\description{
bstrobj <-> data.frame
}
\examples{
test <- bstr_rand_seq(3, 10, seed = 1)

test \%>\% bstr2df
test \%>\% bstr2longdf

test \%>\% bstr2df \%>\% df2bstr
test \%>\% bstr2longdf \%>\% longdf2bstr

}
