% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-bstr.R
\name{bstr_add_attr_for_seq}
\alias{bstr_add_attr_for_seq}
\title{Add sequence attribute}
\usage{
bstr_add_attr_for_seq(bstrobj, attr_name, attrs)
}
\arguments{
\item{bstrobj}{bstring object}

\item{attr_name}{A character. attribute names}

\item{attrs}{attrs}
}
\description{
Add sequence attribute
}
