% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-bstr.R
\name{bstr_add_obj_at}
\alias{bstr_add_obj_at}
\title{Add object attribute}
\usage{
bstr_add_obj_at(bstrobj, attr_name, attrs)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}

\item{attr_name}{one string. an attribute name.}

\item{attrs}{object attribute}
}
\description{
Add object attribute
}
\examples{
temp <- bstr("hogehogehoge")
temp <- temp \%>\% bstr_add_obj_at("hige", "hage")
temp
str(temp)
temp \%>\% bstr_add_obj_at("hige", NA) \%>\% str

}
