% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-bstr.R
\name{bstr_add_seq_at}
\alias{bstr_add_seq_at}
\title{Add sequence attribute}
\usage{
bstr_add_seq_at(bstrobj, attr_name, attrs)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}

\item{attr_name}{one string. an attribute name.}

\item{attrs}{sequence attributes. its length must equal to the length object}
}
\description{
Add sequence attribute
}
\examples{
temp <- dstr_rand_seq(3, 200, seed = 1) \%>\%
  bstr_add_seq_at("circle", c(TRUE, FALSE, TRUE))
temp
temp \%>\% str
# dstr_rand_seq(3, 200) \%>\% bstr_add_seq_at("circle", TRUE)
temp \%>\% bstr_add_seq_at("circle", c(FALSE, TRUE, TRUE))

}
