% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr_count.R
\name{bstr_count}
\alias{bstr_count}
\title{Count the number of matches in a bstr object}
\usage{
bstr_count(bstrobj, pattern = "")
}
\arguments{
\item{bstrobj}{bstr class object or character vector}

\item{pattern}{regex pattern}
}
\description{
Count the number of matches in a bstr object
}
\examples{
bstr_count(c("apple", "banana"))
bstr_count(c("apple", "banana"), "a")
bstr_count(c("apple", "banana"), c("a", "n"))

}
