% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr-methods.R
\name{bstr_degap}
\alias{bstr_degap}
\title{Remove gap character}
\usage{
bstr_degap(bstrobj, gap_chr = "-")
}
\arguments{
\item{bstrobj}{bstr class object or character vector}

\item{gap_chr}{a gap character}
}
\description{
Remove gap character
}
