% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr_detect.R
\name{bstr_detect}
\alias{bstr_detect}
\title{Detect the presence or absence of a pattern in a bstr object}
\usage{
bstr_detect(bstrobj, pattern, negate = FALSE)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}

\item{pattern}{regex pattern}

\item{negate}{If TRUE, return non-matching elements.}
}
\description{
Detect the presence or absence of a pattern in a bstr object
}
\examples{
bstr_detect(c("apple", "banana"), "n")
bstr_detect(c("apple", "banana"), c("a", "n"))

}
