% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr-methods.R
\name{bstr_length}
\alias{bstr_length}
\title{Sequence length in the bstr sequences.}
\usage{
bstr_length(bstrobj)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}
}
\description{
Sequence length in the bstr sequences.
}
\examples{
bstr_length(c("ATG", "cc tgAGT2--"))
bstr_length(NA_character_)

}
