% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr_locate.R
\name{bstr_locate}
\alias{bstr_locate}
\title{Locate all positions of patterns in a bstr object}
\usage{
bstr_locate(bstrobj, pattern)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}

\item{pattern}{regex pattern}
}
\description{
Locate all positions of patterns in a bstr object
}
\examples{
bstr_locate(c("ABCDE", "BCDE"), "A")
bstr_locate(c("ABCDE"), c("A", "C"))
bstr_locate(c("ABCDE", "BC"), c("A", "D", "F"))

}
