% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{bstr_multi_align}
\alias{bstr_multi_align}
\title{Align multiple bstr sequence}
\usage{
bstr_multi_align(subject, queries)
}
\arguments{
\item{subject}{subject sequence}

\item{queries}{query sequences}
}
\description{
Align multiple bstr sequence
}
