% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_seq.R
\name{bstr_rand_seq}
\alias{bstr_rand_seq}
\title{Generate random bstr sequences of desired lengths.}
\usage{
bstr_rand_seq(n, length, pattern = "[A-Za-z]")
}
\arguments{
\item{n}{single integer, nubmer of sequences}

\item{length}{integer vector, desired sequence lengths}

\item{pattern}{character vector specifying character classes}
}
\description{
Generate random bstr sequences of desired lengths.
}
\examples{
bstr_rand_seq(3, 20)
bstr_rand_seq(3, 5:7, c("[a-c]", "[A-C]", "[1-3]"))

}
