% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr-methods-remove.R
\name{bstr_remove}
\alias{bstr_remove}
\title{bstr_remove}
\usage{
bstr_remove(bstrobj, pattern, case_sensitive = F)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}

\item{pattern}{regex pattern}

\item{case_sensitive}{sensitive to case in pattern (default:FALSE)}
}
\description{
bstr_remove
}
