% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr-methods.R
\name{bstr_remove_all}
\alias{bstr_remove_all}
\title{bstr_remove_all}
\usage{
bstr_remove_all(bstrobj, pattern)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}

\item{pattern}{regex pattern}
}
\description{
bstr_remove_all
}
