% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr-methods-remove.R
\name{bstr_remove_gap}
\alias{bstr_remove_gap}
\title{Remove all gap character}
\usage{
bstr_remove_gap(bstrobj, gap_chr = "-")
}
\arguments{
\item{bstrobj}{bstr class object or character vector}

\item{gap_chr}{a gap character}
}
\description{
Remove all gap character
}
