% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr-methods-remove.R
\name{bstr_remove_notalpha}
\alias{bstr_remove_notalpha}
\title{bstr_remove_notalpha}
\usage{
bstr_remove_notalpha(bstrobj)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}
}
\description{
bstr_remove_notalpha
}
