% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr-methods.R
\name{bstr_replace}
\alias{bstr_replace}
\title{bstr_replace}
\usage{
bstr_replace(bstrobj, pattern, replacement)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}

\item{pattern}{regex pattern}

\item{replacement}{a character vector}
}
\description{
bstr_replace
}
