% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{bstr_replace_name}
\alias{bstr_replace_name}
\title{Replace names of the bstring object}
\usage{
bstr_replace_name(bstrobj, pattern, replacement)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}

\item{pattern}{Pattern to look for.}

\item{replacement}{A character vector of replacements or a function.}
}
\description{
Replace names of the bstring object
}
\examples{
temp <- bstr(c("aaa", "bbb"))
c(temp, bstr_replace_name(temp, "\\\\d", c("one", "two")))

}
