% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr-methods-trans.R
\name{bstr_reverse}
\alias{bstr_reverse}
\title{Reverse bstr sequence}
\usage{
bstr_reverse(bstrobj)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}
}
\description{
Reverse bstr sequence
}
\examples{
temp <- bstr_rand_seq(2, 6)
c(temp, reverse = bstr_reverse(temp))

}
