% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr-methods.R
\name{bstr_sub}
\alias{bstr_sub}
\title{bstr_sub}
\usage{
bstr_sub(bstrobj, start = 1L, end = -1L)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}

\item{start}{start}

\item{end}{end}
}
\description{
bstr_sub
}
