% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr-methods.R
\name{bstr_sub_true}
\alias{bstr_sub_true}
\title{bstr_sub_true}
\usage{
bstr_sub_true(bstrobj, i, start, end, gap_chr = "-")
}
\arguments{
\item{bstrobj}{bstr class object or character vector}

\item{i}{index}

\item{start}{start}

\item{end}{end}

\item{gap_chr}{gap character}
}
\description{
bstr_sub_true
}
