% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr-methods-trans.R
\name{bstr_switch_case}
\alias{bstr_switch_case}
\title{bstr_switch_case}
\usage{
bstr_switch_case(bstrobj)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}
}
\description{
bstr_switch_case
}
