% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_identity.R
\name{calc_ident}
\alias{calc_ident}
\title{calculate \% of identical residues}
\usage{
calc_ident(x, y, gap_chr = "-")
}
\arguments{
\item{x}{aligned sequence 1}

\item{y}{aligned sequence 2}

\item{gap_chr}{gap character}
}
\description{
calculate \% of identical residues
}
