% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_identity.R
\name{calc_ident_mat}
\alias{calc_ident_mat}
\title{calculate \% of identical residues}
\usage{
calc_ident_mat(bstrobj, gap_chr = "-")
}
\arguments{
\item{bstrobj}{bstrobj}

\item{gap_chr}{gap character}
}
\description{
calculate \% of identical residues
}
\examples{
temp <- bstr(c("ATGC--CG", "ACGC-GGG", "--------", "AT"))
round(calc_ident_mat(temp), 2)
}
