% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infusion.R
\name{calc_InFusion_vector_volume}
\alias{calc_InFusion_vector_volume}
\title{Calculate the vector volume for In-Fusion reaction}
\usage{
calc_InFusion_vector_volume(
  seq_i,
  seq_v,
  conc_i,
  conc_v,
  uL_i,
  ratio_i = 2,
  ratio_v = 1
)
}
\arguments{
\item{seq_i}{DNA sequence of insertion}

\item{seq_v}{DNA sequence of vector}

\item{conc_i}{DNA concentration (ng/uL) of purified insertion}

\item{conc_v}{DNA concentration (ng/uL) of purified vector}

\item{uL_i}{Supplied volume (uL) of purified insertion}

\item{ratio_i}{Molecular ratio of insertion}

\item{ratio_v}{Molecular ratio of vector}
}
\description{
Calculate the vector volume for In-Fusion reaction
}
\examples{
calc_InFusion_vector_volume(
  seq_i = dstr_rand_seq(1, 1000, seed = 1),
  seq_v = dstr_rand_seq(1, 5000, seed = 1),
  conc_i = 5, conc_v = 30,
  uL_i = 3
)

}
