% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mw.R
\name{calc_mw}
\alias{calc_mw}
\alias{calc_mw_dsDNA}
\alias{calc_mw_ssRNA}
\alias{calc_mw_peptide}
\title{calculate molecular weight}
\usage{
calc_mw_dsDNA(dstrobj)

calc_mw_ssRNA(dstrobj)

calc_mw_peptide(pstrobj)
}
\arguments{
\item{dstrobj}{dstr class object or character vector}

\item{pstrobj}{pstr class object or character vector}
}
\description{
calculate molecular weight

Calculate double strand DNA molecular weight

Calculate single strand RNA molecular weight

Calculate peptide molecular weight
}
\examples{

(test <- dstr_rand_seq(3, 20, seed = 1))
calc_mw_dsDNA(test)
calc_mw_ssRNA(test)

(res <- pstr_rand_seq(3, 100, seed = 1) \%>\% calc_mw_peptide())
barplot(percent ~ name + aa, legend.text = unique(res$name),
  beside = TRUE, data = res)

}
