% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mw.R
\name{calc_MW_dsDNA}
\alias{calc_MW_dsDNA}
\title{Calculate double strand DNA molecular weight}
\usage{
calc_MW_dsDNA(dna)
}
\arguments{
\item{dna}{dna sequence}
}
\description{
Calculate double strand DNA molecular weight
}
