% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mw.R
\name{calc_MW_peptide}
\alias{calc_MW_peptide}
\title{Calculate protein molecular weight}
\usage{
calc_MW_peptide(astrobj)
}
\arguments{
\item{astrobj}{astr object.}
}
\description{
Calculate protein molecular weight
}
\examples{
astr_rand_seq(3, 100) \%>\% calc_MW_peptide()

}
