% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein.R
\name{calc_MW_ssRNA}
\alias{calc_MW_ssRNA}
\title{Calculate single strand RNA molecular weight}
\usage{
calc_MW_ssRNA(dna)
}
\arguments{
\item{dna}{dna sequence}
}
\description{
Calculate single strand RNA molecular weight
}
