% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr_methods-primer.R
\name{calc_oligoDNATm}
\alias{calc_oligoDNATm}
\title{Calculate Oligo DNA Tm}
\usage{
calc_oligoDNATm(x)
}
\arguments{
\item{x}{A DNA sequence you want to calculate melting temperature}
}
\description{
Calculate Oligo DNA Tm
}
