% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr-methods-trans.R
\name{case}
\alias{case}
\alias{bstr_to_upper}
\alias{bstr_to_lower}
\alias{bstr_switch_case}
\title{Convert case of a bstring sequence.}
\usage{
bstr_to_upper(bstrobj)

bstr_to_lower(bstrobj)

bstr_switch_case(bstrobj)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}
}
\description{
Convert case of a bstring sequence.
}
\examples{
temp <- dstr_rand_seq(1, 10, "[atgcATGC]")
c(
  temp,
  upper = bstr_to_upper(temp),
  lower = bstr_to_lower(temp),
  switch = bstr_switch_case(temp)
)

}
