% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-bstr.R
\name{class_bstr}
\alias{class_bstr}
\title{Common bstr class augments}
\arguments{
\item{x}{A character vector which convert to a bstr object.}

\item{n}{A character vector which is name of x.}

\item{ucase}{A logical. If TRUE the x is converted to upper case. (default: FALSE)}

\item{bstrobj}{bstr class object or character vector}

\item{dstrobj}{dstr class object or character vector}

\item{astrobj}{astr class object or character vector}

\item{pattern}{regex pattern}

\item{case_sensitive}{sensitive to case in pattern (default:FALSE)}
}
\description{
Common bstr class augments
}
