% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr-methods.R
\name{class_bstr_arg}
\alias{class_bstr_arg}
\title{Common bstr class augments}
\usage{
class_bstr_arg(bstrobj, pattern)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}

\item{pattern}{regex pattern}
}
\description{
Common bstr class augments
}
