% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dstr.R
\name{dstr}
\alias{dstr}
\title{Constructer of the dstr class object}
\usage{
dstr(x, n, ucase = FALSE)
}
\arguments{
\item{x}{A character vector which convert to a bstr object.}

\item{n}{A character vector which is name of x.}

\item{ucase}{A logical. If TRUE the x is converted to upper case. (default: FALSE)}
}
\description{
Constructer of the dstr class object
}
\examples{
dstr("A.Bad.bat", "It is")
dstr(c("A", "bad", "Bat"), ucase = TRUE)

}
