% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{dstr_align}
\alias{dstr_align}
\title{Align dna sequence}
\usage{
dstr_align(seq1, seq2, rc = F)
}
\arguments{
\item{seq1}{sequence}

\item{seq2}{sequence}

\item{rc}{logical value. if TRUE, the seq2 is aligned after reverse complemented. default is FALSE.}
}
\description{
Align dna sequence
}
