% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr_misc_both.R
\name{dstr_both}
\alias{dstr_both}
\alias{dstr_detect_both}
\alias{dstr_count_both}
\alias{dstr_locate_both}
\title{Apply function to both forward and reverse-complement sequences.}
\usage{
dstr_detect_both(dstrobj, pattern, negate = FALSE)

dstr_count_both(dstrobj, pattern)

dstr_locate_both(dstrobj, pattern)
}
\arguments{
\item{dstrobj}{dstr class object or character vector}

\item{pattern}{regex pattern}

\item{negate}{If TRUE, return non-matching elements.}
}
\description{
Apply function to both forward and reverse-complement sequences.
}
\examples{
temp <- c("aaagggcccc", "aaagggtttcccctttccc", "gggtttccc")
dstr_detect_both(temp, "aaa")
dstr_count_both(temp, "aaa")
dstr_locate_both(temp, "aaa")
dstr_locate_both(temp, "aaa") \%>\%
  lapply(. \%>\% bstr_sub_all(temp, .))

}
