% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr-methods.R
\name{dstr_complement_fast}
\alias{dstr_complement_fast}
\title{Fast version of dstr_complement with no case conservation}
\usage{
dstr_complement_fast(dstrobj)
}
\arguments{
\item{dstrobj}{dstr object}
}
\description{
Fast version of dstr_complement with no case conservation
}
