% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr-methods.R
\name{dstr_convert_iupac2regex}
\alias{dstr_convert_iupac2regex}
\title{convert IUPAC CODE to regular expression}
\usage{
dstr_convert_iupac2regex(dna)
}
\arguments{
\item{dna}{A character}
}
\description{
convert IUPAC CODE to regular expression
}
