% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr_misc_gw.R
\name{dstr_extract_orfs_from_vector}
\alias{dstr_extract_orfs_from_vector}
\alias{dstr_find_orfs_from_vector}
\alias{dstr_translate_from_vector}
\title{Find open reading frames from Gateway system compatible vector}
\usage{
dstr_extract_orfs_from_vector(dstrobj)

dstr_find_orfs_from_vector(dstrobj)

dstr_translate_from_vector(dstrobj)
}
\arguments{
\item{dstrobj}{dstr class object or character vector}
}
\description{
Find open reading frames from Gateway system compatible vector
}
\examples{
attR1 <- "ACAAGTTTGTACAAAAAAGCTGAAC"
attR2 <- "GTTCAGCTTTCTTGTACAAAGTGGT"
A9 <- "AAAAAAAAA"
dummy <- paste0(A9, attR1, A9, attR2, A9, collapse = "")
dummy2 <- dstr_rev_comp(dummy)
dummy3 <- paste0(dummy, dummy2, collapse = "")
test_distination <- dstr(c(dummy, dummy2, dummy3), paste0("dummy", 1:3))

vectors <-
  dstr_gw_topo("atgtga") \%>\%
  dstr_gw_lr_both(test_distination)
vectors

dstr_extract_orfs_from_vector(vectors)
dstr_translate_from_vector(vectors)
dstr_translate_from_vector(dstr_rev_comp_fast(vectors))

}
