% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr-methods.R
\name{dstr_find_orf}
\alias{dstr_find_orf}
\title{Find open reading frames from DNA}
\usage{
dstr_find_orf(dstrobj, search_none_stop = F)
}
\arguments{
\item{dstrobj}{a sequence you want to find orfs}

\item{search_none_stop}{A logical. If TRUE, only search orfs with a stop codon. defalut is FALSE}
}
\description{
Find open reading frames from DNA
}
