% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr-methods-gateway_system.R
\name{dstr_find_orf_from_vector}
\alias{dstr_find_orf_from_vector}
\title{Find a main open reading frame from Gataway system vector}
\usage{
dstr_find_orf_from_vector(x)
}
\arguments{
\item{x}{a vector sequence you want to find orfs}
}
\description{
Find a main open reading frame from Gataway system vector
}
