% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr-methods-gateway_system.R
\name{dstr_gw_lr}
\alias{dstr_gw_lr}
\title{Return product sequence after LR reaction}
\usage{
dstr_gw_lr(entry, distination)
}
\arguments{
\item{entry}{Entry clone}

\item{distination}{Distination vector}
}
\description{
Return product sequence after LR reaction
}
