% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr-methods-gateway_system.R
\name{dstr_gw_topo}
\alias{dstr_gw_topo}
\title{Return the product sequence of after TOPO reaction}
\usage{
dstr_gw_topo(insert, pentr)
}
\arguments{
\item{insert}{insert sequence. without 5'CACC}

\item{pentr}{optional. pENTR D-TOPO sequence or the fasta file path}
}
\description{
Return the product sequence of after TOPO reaction
}
\examples{
dstr_gw_topo("atgatgatgtga")

}
