% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr-methods-gateway_system.R
\name{dstr_gw_topo_lr}
\alias{dstr_gw_topo_lr}
\title{TOPO and LR reaction}
\usage{
dstr_gw_topo_lr(insert, distination, pentr)
}
\arguments{
\item{insert}{insert sequence. without 5'CACC}

\item{distination}{Distination vector}

\item{pentr}{optional. pENTR D-TOPO sequence or the fasta file path}
}
\description{
TOPO and LR reaction
}
