% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr-methods.R
\name{dstr_iupac2regex}
\alias{dstr_iupac2regex}
\title{Convert IUPAC CODE to regular expression}
\usage{
dstr_iupac2regex(dstrobj)
}
\arguments{
\item{dstrobj}{A character}
}
\description{
Convert IUPAC CODE to regular expression
}
\examples{
dstr_iupac2regex("MRWSYKVHDBNNN")

}
