% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{dstr_pcr}
\alias{dstr_pcr}
\title{Calculate PCR product length}
\usage{
dstr_pcr(template, primerF, primerR, FRC = FALSE, RRC = TRUE)
}
\arguments{
\item{template}{template DNA sequence}

\item{primerF}{left primer sequence}

\item{primerR}{right primer sequence}

\item{FRC}{logical. TRUE when reverse complemented primer F. default is FALSE.}

\item{RRC}{logical. TRUE when reverse complemented primer R. default is TRUE.}
}
\description{
Calculate PCR product length
}
\examples{
dstr_pcr("ACAATGTGTGTATGATGGTAGTAGAC", "ATGTG", "TACTA")
dstr_pcr("ACAATGTGTGTATGATGGTAGTAGAC", "ATGTG", "TACTA") \%>\%
  {dstr_align_multi(.[1], .[-1])}

}
