% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_seq.R
\name{dstr_rand_seq}
\alias{dstr_rand_seq}
\title{Generate random dstr sequences of desired lengths.}
\usage{
dstr_rand_seq(n, length, pattern = "[ATGC]", seed)
}
\arguments{
\item{n}{single integer, nubmer of sequences}

\item{length}{integer vector, desired sequence lengths}

\item{pattern}{character vector specifying character classes}

\item{seed}{a single integer. specify seeds.}
}
\description{
Generate random dstr sequences of desired lengths.
}
\examples{
dstr_rand_seq(3, 20, seed = 1)
dstr_rand_seq(3, 5:7, c("[atcg]", "[ATCG]", "[N.]"), seed = 1)

}
