% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr-methods.R
\name{dstr_remove_all}
\alias{dstr_remove_all}
\title{dstr_remove_all}
\usage{
dstr_remove_all(dstrobj, pattern)
}
\arguments{
\item{dstrobj}{dstr class object or character vector}

\item{pattern}{regex pattern}
}
\description{
dstr_remove_all
}
