% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr-methods-gateway_system.R
\name{dstr_TOPO}
\alias{dstr_TOPO}
\title{Return the product sequence of after TOPO reaction}
\usage{
dstr_TOPO(cds)
}
\arguments{
\item{cds}{coding sequence which you want to clone. without 5'CACC}
}
\description{
Return the product sequence of after TOPO reaction
}
