% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr-methods-gateway_system.R
\name{dstr_TOPO_LR}
\alias{dstr_TOPO_LR}
\title{TOPO and LR reaction}
\usage{
dstr_TOPO_LR(cds, distination)
}
\arguments{
\item{cds}{coding sequence which you want to clone. without 5'CACC}

\item{distination}{Distination vector}
}
\description{
TOPO and LR reaction
}
