% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr_translate.R
\name{dstr_translate}
\alias{dstr_translate}
\title{Translate dna -> protein}
\usage{
dstr_translate(dstrobj)
}
\arguments{
\item{dstrobj}{sequence you want to translate}
}
\description{
Translate dna -> protein
}
