% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr-methods-gateway_system.R
\name{dstr_translate_from_vector}
\alias{dstr_translate_from_vector}
\title{Find and translate ORF from expression vector}
\usage{
dstr_translate_from_vector(dna)
}
\arguments{
\item{dna}{sequence you want to translate}
}
\description{
Find and translate ORF from expression vector
}
