% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr-methods.R
\name{dstr_trim_stop}
\alias{dstr_trim_stop}
\title{Remove stop codon from DNA sequence}
\usage{
dstr_trim_stop(dstrobj, stop_codon = "(TAA|TGA|TAG)$")
}
\arguments{
\item{dstrobj}{dstr class object or character vector}

\item{stop_codon}{stop codon pattern}
}
\description{
Remove stop codon from DNA sequence
}
