% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-bstr.R
\name{format_parts_name}
\alias{format_parts_name}
\title{format bstr name}
\usage{
format_parts_name(x, l_name)
}
\arguments{
\item{x}{x}

\item{l_name}{max length of name}
}
\description{
format bstr name
}
