% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-bstr.R
\name{format_row}
\alias{format_row}
\title{format bstr}
\usage{
format_row(x, n, l_name, l_seq)
}
\arguments{
\item{x}{x}

\item{n}{number of printing elements}

\item{l_name}{max length of name}

\item{l_seq}{max length of sequence}
}
\description{
format bstr
}
