% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-astr.R
\name{is_valid_aa_character}
\alias{is_valid_aa_character}
\title{Detect invalid IUPAC Amino Acid characters}
\usage{
is_valid_aa_character(x, negate = FALSE)
}
\arguments{
\item{x}{A character vector which convert to a bstr object.}

\item{negate}{If TRUE, return non-matching elements.}
}
\description{
Detect invalid IUPAC Amino Acid characters
}
\examples{
test <- c(letters[1:10])
test2 <-
  c(paste(test[1:5], collapse = ""), paste(test[6:10], collapse = ""))
is_valid_aa_character(test) \%>\% table
is_valid_aa_character(test, negate = TRUE) \%>\% test[.]
is_valid_aa_character(test2)

}
