% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dstr.R
\name{is_valid_dna_character}
\alias{is_valid_dna_character}
\title{Detect invalid IUPAC DNA characters}
\usage{
is_valid_dna_character(x, negate = FALSE)
}
\arguments{
\item{x}{A character vector which convert to a bstr object.}
}
\description{
Detect invalid IUPAC DNA characters
}
