% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr_misc_orfs.R
\name{is_valid_orf}
\alias{is_valid_orf}
\title{Check open reading frame validity}
\usage{
is_valid_orf(dstrobj, check_stop = FALSE, negate = FALSE)
}
\arguments{
\item{dstrobj}{dstr class object or character vector}

\item{check_stop}{If TRUE, check the presence of stop codon.}

\item{negate}{If TRUE, return non-matching elements.}
}
\description{
Check open reading frame validity
}
\examples{
(test <- dstr(c("ATGtga", "AAGTGA", "ATGGGGTGA", "ATGGGTGA", "ATGGGG","ATGGG")))
is_valid_orf(test)
is_valid_orf(test, check_stop = TRUE)

}
