% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr_length.R
\name{length}
\alias{length}
\alias{bstr_length}
\alias{dstr_length}
\alias{astr_length}
\title{Sequence length in the bstr sequences.}
\usage{
bstr_length(bstrobj)

dstr_length(bstrobj)

astr_length(bstrobj)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}
}
\description{
Sequence length in the bstr sequences.
}
\examples{
test <- c("ATG", "cc tgAGT2--")
bstr_length(test)
bstr_length(NA_character_)

}
