% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr_methods-primer.R
\name{mf}
\alias{mf}
\title{Calculate Oligo DNA Tm}
\usage{
mf(numTotal, numA, numT, numG, numC, dinuc1, dinuc2)
}
\arguments{
\item{numTotal}{hoge}

\item{numA}{hoge}

\item{numT}{hoge}

\item{numG}{hoge}

\item{numC}{hoge}

\item{dinuc1}{hoge}

\item{dinuc2}{hoge}
}
\description{
Calculate Oligo DNA Tm
}
