% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr_misc_orfs.R
\name{orfs}
\alias{orfs}
\alias{dstr_locate_orfs}
\alias{dstr_extract_orfs}
\alias{dstr_find_orfs}
\title{Find and extract all open reading frames in a dstr object}
\usage{
dstr_locate_orfs(dstrobj, orf_pattern = REGEX_ORF[2])

dstr_extract_orfs(dstrobj, orf_pattern = REGEX_ORF[2])

dstr_find_orfs(dstrobj, orf_pattern = REGEX_ORF[2])
}
\arguments{
\item{dstrobj}{dstr class object or character vector}

\item{orf_pattern}{regular expression pattern of open reading frame}
}
\description{
Find and extract all open reading frames in a dstr object
}
\examples{
(temp <- dstr_rand_seq(2, 200, seed = 1))
temp \%>\% dstr_locate_orfs() \%>\% {bstr_sub_all(temp, .)}
dstr_locate_orfs(c("atgatgtga", "aaa"))

(temp <- dstr_rand_seq(2, 200, seed = 1))
temp \%>\% dstr_extract_orfs()
# temp \%>\% dstr_extract_orfs(REGEX_ORF)

}
