% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_genbank.R
\name{parse_DEFINITION_FIELD}
\alias{parse_DEFINITION_FIELD}
\title{Parse Genbank file DIFINITION FIELD}
\usage{
parse_DEFINITION_FIELD(gb_line)
}
\arguments{
\item{gb_line}{character vector of a Genbank file.}
}
\description{
Parse Genbank file DIFINITION FIELD
}
\examples{
inf <- system.file("extdata", "sample.gb", package = "bstringr")
readLines(inf) \%>\% parse_DEFINITION_FIELD()

}
