% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_genbank.R
\name{parse_LOCUS_FIELD}
\alias{parse_LOCUS_FIELD}
\title{Parse Genbank file LOCUS FIELD}
\usage{
parse_LOCUS_FIELD(gb_line)
}
\arguments{
\item{gb_line}{character vector of a Genbank file.}
}
\description{
Parse Genbank file LOCUS FIELD
}
\examples{
inf <- system.file("extdata", "sample.gb", package = "bstringr")
readLines(inf) \%>\% parse_LOCUS_FIELD()

}
