% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_genbank.R
\name{parse_ORIGIN_FIELD}
\alias{parse_ORIGIN_FIELD}
\title{Parse Genbank file ORIGIN FIELD}
\usage{
parse_ORIGIN_FIELD(gb_line)
}
\arguments{
\item{gb_line}{character vector of a Genbank file.}
}
\description{
Parse Genbank file ORIGIN FIELD
}
\examples{
inf <- system.file("extdata", "sample.gb", package = "bstringr")
readLines(inf) \%>\% parse_ORIGIN_FIELD()

}
