% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_genbank.R
\name{parse_VERSION_FIELD}
\alias{parse_VERSION_FIELD}
\title{Parse Genbank file VERSION FIELD}
\usage{
parse_VERSION_FIELD(gb_line)
}
\arguments{
\item{gb_line}{character vector of a Genbank file.}
}
\description{
Parse Genbank file VERSION FIELD
}
\examples{
inf <- system.file("extdata", "sample.gb", package = "bstringr")
readLines(inf) \%>\% parse_VERSION_FIELD()

}
