% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-bstr.R
\name{print.bstr}
\alias{print.bstr}
\title{print bstr class object}
\usage{
\method{print}{bstr}(x, ..., n = 6L, l_name = 20L, l_seq = 40L)
}
\arguments{
\item{x}{x}

\item{...}{...}

\item{n}{number of printing elements}

\item{l_name}{max length of name}

\item{l_seq}{max length of sequence}
}
\description{
print bstr class object
}
\examples{
dstr_rand_seq(10, 20, seed = 1)
pstr_rand_seq(10, 20, seed = 1)

test1 <- dstr_rand_seq(1, 20, seed = 1)
test2 <- dstr_rand_seq(1, 21, seed = 1)
print(test1, l_seq = 20L)
print(test2, l_seq = 20L)

print(test1, l_name = 9L)
print(test1, l_name = 8L)

dstr_rand_seq(10, 20, seed = 1) \%>\% print(n = 20L)

}
