% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_fasta.R
\name{read_fasta}
\alias{read_fasta}
\title{Read fasta file}
\usage{
read_fasta(fpath, ucase = F)
}
\arguments{
\item{fpath}{A fasta file path.}

\item{ucase}{logical}
}
\description{
Read fasta file
}
\examples{
inf <- system.file("extdata", "test.fas", package = "bstringr")
read_fasta(inf)

}
