% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_genbank.R
\name{read_genbank}
\alias{read_genbank}
\title{Read Genbank file}
\usage{
read_genbank(fpath, ucase = F)
}
\arguments{
\item{fpath}{A genbank file path.}

\item{ucase}{logical}
}
\description{
Read Genbank file
}
\examples{
inf <- system.file("extdata", "sample.gb", package = "bstringr")
read_genbank(inf)

}
