% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr_reverse.R
\name{reverse}
\alias{reverse}
\alias{bstr_reverse}
\alias{dstr_reverse}
\alias{astr_reverse}
\title{Reverse bstr sequence}
\usage{
bstr_reverse(bstrobj)

dstr_reverse(bstrobj)

astr_reverse(bstrobj)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}
}
\description{
Reverse bstr sequence
}
\examples{
temp <- bstr_rand_seq(2, 6, seed = 1)
c(temp, reverse = bstr_reverse(temp))

}
