% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_align.R
\name{show_align}
\alias{show_align}
\title{show alignment}
\usage{
show_align(bstrobj, length, gap_chr = "-")
}
\arguments{
\item{bstrobj}{bstr class object or character vector}

\item{length}{column length}

\item{gap_chr}{a gap character}
}
\description{
show alignment
}
\examples{
dstr_rand_seq(5, seq(40, 110, l = 5), seed = 1) \%>\%
  show_align(length = 40)

bstr_rand_seq(2, c(15, 30), "[AT]", seed = 2) \%>\%
  {bstr_align_pairwise(.[1], .[2])} \%>\%
  show_align(length = 40)

}
