% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{sort.bstr}
\alias{sort.bstr}
\title{sort bstr}
\usage{
\method{sort}{bstr}(x, decreasing = FALSE, ..., by = c("names", "length"))
}
\arguments{
\item{x}{x}

\item{decreasing}{logical. FALSE}

\item{...}{...}

\item{by}{c("names", "length")}
}
\description{
sort bstr
}
\examples{
test <- bstr_rand_seq(6, c(10, 5, 10, 12, 13, 3), seed = 1)
test
sort(test, decreasing = TRUE)
sort(test, by = "length")

}
