% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bstr.R
\name{split_sequence}
\alias{split_sequence}
\title{Split sequence by residues and return data.frame}
\usage{
split_sequence(seq, seq_name_fct)
}
\arguments{
\item{seq}{A character vector.}

\item{seq_name_fct}{A factor.}
}
\description{
Split sequence by residues and return data.frame
}
