% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{[.bstr}
\alias{[.bstr}
\title{subsetting biostrings class object}
\usage{
\method{[}{bstr}(x, ..., drop = F)
}
\arguments{
\item{x}{x}

\item{...}{...}

\item{drop}{drop}
}
\description{
subsetting biostrings class object
}
