% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr_sub.R
\name{sub}
\alias{sub}
\alias{bstr_sub}
\alias{bstr_sub<-}
\alias{bstr_sub_replace}
\title{Extract and replace subsequences from a bstr sequences}
\usage{
bstr_sub(bstrobj, start = 1L, end = -1L)

bstr_sub(bstrobj, start = 1L, end = -1L, omit_na = FALSE) <- value

bstr_sub_replace(..., replacement, value = replacement)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}

\item{start}{start}

\item{end}{end}

\item{omit_na}{Single logical value. If `TRUE`, missing values in any of the
arguments provided will result in an unchanged input.}

\item{value}{replacement string}

\item{...}{arguments to be passed to \code{bstr_sub<-}}

\item{replacement}{alias of \code{value} [wherever applicable]}
}
\description{
Extract and replace subsequences from a bstr sequences
}
\examples{
temp <- bstr_rand_seq(1, 10, seed = 1)
c(
  temp,
  bstr_sub(temp, 1, 5),
  bstr_sub(temp, 5),
  bstr_sub(temp, end = 5),
  bstr_sub(temp, 1:2, c(6,2))
)

# Negative indices
c(
  temp,
  bstr_sub(temp, -1),
  bstr_sub(temp, -7),
  bstr_sub(temp, end = -7)
)

# Alternatively, you can pass in a two colum matrix, as in the
# output from bstr_locate
temp <- bstr("abcde fghij")
pos <- bstr_locate(temp, "[aefi]")[[1]]
bstr_sub(temp, pos)

# Vectorisation
bstr_sub(temp, seq_len(bstr_length(temp)))
bstr_sub(temp, end = seq_len(bstr_length(temp)))

# Replacement form
x <- "BBCDEF"
bstr_sub(x, 1, 1) <- "A"; x
bstr_sub(x, -1, -1) <- "K"; x
bstr_sub(x, -2, -2) <- "GHIJ"; x
bstr_sub(x, 2, -2) <- ""; x

# If you want to keep the original if some argument is NA,
# use omit_na = TRUE
x1 <- x2 <- x3 <- x4 <- bstr("AAA")
bstr_sub(x1, 1, NA) <- "B"
bstr_sub(x2, 1, 2) <- NA
bstr_sub(x3, 1, NA, omit_na = TRUE) <- "B"
bstr_sub(x4, 1, 2, omit_na = TRUE) <- NA
c(x1, x2, x3, x4)

}
