% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_fasta.R
\name{write_fasta}
\alias{write_fasta}
\title{write fasta file}
\usage{
write_fasta(x, fpath, width = 60)
}
\arguments{
\item{x}{x}

\item{fpath}{A fasta file path.}

\item{width}{w}
}
\description{
write fasta file
}
\examples{
(temp <- dstr_rand_seq(n = 3, length = c(60, 90, 200), seed = 1))
write_fasta(temp)

}
