% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_genbank.R
\name{write_genbank}
\alias{write_genbank}
\title{Write Genbank file}
\usage{
write_genbank(x, fpath)
}
\arguments{
\item{x}{x}

\item{fpath}{A genbank file path.}
}
\description{
Write Genbank file
}
\examples{
(temp <- dstr_rand_seq(1, 100, seed = 1))
write_genbank(temp)

}
