% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align_multi}
\alias{align_multi}
\alias{bstr_align_multi}
\alias{dstr_align_multi}
\title{Align multiple sequences to reference}
\usage{
bstr_align_multi(sub, pat, type = "global", gapO = 10, gapE = 4)

dstr_align_multi(sub, pat, rc, type = "global", gapO = 10, gapE = 4)
}
\arguments{
\item{sub}{a subject sequence}

\item{pat}{pattern sequence}

\item{type}{type of alignment. One of "global", "local", "overlap",
"global-local", and "local-global".}

\item{gapO}{gap opening penalty}

\item{gapE}{gap extension penalty}

\item{rc}{logical value. if TRUE, the seq2 is aligned after reverse complemented. default is FALSE.}
}
\description{
Align multiple sequences to reference
}
\examples{
ref <- c(reference = "abcdefghijk")
patterns <- c(one = "abc", two = "d", three = "efgjk")
bstr_align_multi(ref, patterns)

ref <- c(ref = "AAACCCTTTGGG")
patterns <- c(A = "AAA", C = "CCC", T = "TTT", G = "GGG")
dstr_align_multi(ref, patterns)
dstr_align_multi(ref, patterns, 1:2)
dstr_align_multi(ref, patterns, c("T", "G"))

}
