% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-bstr.R
\name{as_bstr}
\alias{as_bstr}
\title{Convert character vector to biostring class}
\usage{
as_bstr(x, n)
}
\arguments{
\item{x}{A character vector which convert to a bstr object.}

\item{n}{A character vector which is name of x.}
}
\description{
Convert character vector to biostring class
}
