% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dstr.R
\name{as_dstr}
\alias{as_dstr}
\title{Convert character vector to dstr class}
\usage{
as_dstr(x, n)
}
\arguments{
\item{x}{A character vector which convert to a bstr object.}

\item{n}{A character vector which is name of x.}
}
\description{
Convert character vector to dstr class
}
