% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dstr.R
\name{as_dstr}
\alias{as_dstr}
\title{Convert character vector to dstr class}
\usage{
as_dstr(x, n, ucase = FALSE)
}
\arguments{
\item{x}{A character vector which convert to a bstr object.}

\item{n}{A character vector which is name of x.}

\item{ucase}{A logical. If TRUE the x is converted to upper case. (default: FALSE)}
}
\description{
Convert character vector to dstr class
}
\examples{
as_dstr("bad", "good", ucase = TRUE)
# as_dstr(c("good", "bad")) # Error

}
