% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-astr.R
\name{astr}
\alias{astr}
\title{Constructing astr class object}
\usage{
astr(x, n, ucase = T)
}
\arguments{
\item{x}{Character vector}

\item{n}{Character vector}

\item{ucase}{Character vector}
}
\description{
Constructing astr class object
}
