% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-bstr.R
\name{bstr}
\alias{bstr}
\title{Constructer of the bstr class object}
\usage{
bstr(x, n, ucase = FALSE)
}
\arguments{
\item{x}{A character vector which convert to a bstr object.}

\item{n}{A character vector which is name of x.}

\item{ucase}{A logical. If TRUE the x is converted to upper case. (default: FALSE)}
}
\description{
Constructer of the bstr class object
}
\examples{
bstr("Apple", "apple")
bstr(c("Apple", "potato"), c("apple", "imo"), ucase = TRUE)

}
