% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bstr.R
\name{bstr2df}
\alias{bstr2df}
\title{Split sequence by residues and return data.frame}
\usage{
bstr2df(bstrobj)
}
\arguments{
\item{bstrobj}{A bstr object or A character vector}
}
\description{
Split sequence by residues and return data.frame
}
\examples{
bstr2df(c("ATGC", "AGCT--GT"))
}
