% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstr-methods-remove.R
\name{bstr_replace}
\alias{bstr_replace}
\title{bstr_replace}
\usage{
bstr_replace(bstrobj, pattern, replacement, case_sensitive = F)
}
\arguments{
\item{bstrobj}{bstr class object or character vector}

\item{pattern}{regex pattern}

\item{replacement}{a character vector}

\item{case_sensitive}{sensitive to case in pattern (default:FALSE)}
}
\description{
bstr_replace
}
