% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{c.bstr}
\alias{c.bstr}
\title{Combine Values into a bstr class object}
\usage{
\method{c}{bstr}(...)
}
\arguments{
\item{...}{...}
}
\description{
Combine Values into a bstr class object
}
\examples{
c(dstr_rand_seq(2, 5, seed = 1), dstr_rand_seq(3, 5, seed = 2))
c(dstr_rand_seq(3, 5, seed = 2), dstr_rand_seq(2, 5, seed = 1))

c("hoge", dstr("atg"))
c(dstr("atg"), "hoge")
c(dstr("atg"), bstr("hoge"))
c(dstr("atg"), dstr("ccc"))
c(dstr("atg"), pstr("ccc"))
c(pstr("atg"), pstr("ccc"))

}
