% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein.R
\name{calc_MW_peptide}
\alias{calc_MW_peptide}
\title{Calculate protein molecular weight}
\usage{
calc_MW_peptide(pro, needinf = c("Sequence", "MW"))
}
\arguments{
\item{pro}{Amino acids sequence you want to calculate molecular weight}

\item{needinf}{information "Sequence", "AAComposition", "AAFrequency", "MW"}
}
\description{
Calculate protein molecular weight
}
