% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dstr.R
\name{dstr}
\alias{dstr}
\title{Constructer of the dstr class object}
\usage{
dstr(x, n, ucase = T)
}
\arguments{
\item{x}{Character vector}

\item{n}{Character vector}

\item{ucase}{Character vector}
}
\description{
Constructer of the dstr class object
}
