% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr-methods.R
\name{dstr_rev_comp}
\alias{dstr_rev_comp}
\title{Return the reverse complement sequence}
\usage{
dstr_rev_comp(dstrobj)
}
\arguments{
\item{dstrobj}{dstr object}
}
\description{
Return the reverse complement sequence
}
