% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr_misc_gw.R
\name{gw}
\alias{gw}
\alias{dstr_gw_topo}
\alias{dstr_gw_lr}
\alias{dstr_gw_topo_lr}
\title{TOPO reaction and LR reaction}
\usage{
dstr_gw_topo(insert, pentr)

dstr_gw_lr(entry, distination)

dstr_gw_topo_lr(insert, distination, pentr)
}
\arguments{
\item{insert}{insert sequence. without 5'CACC}

\item{pentr}{optional. pENTR D-TOPO sequence or the fasta file path}

\item{entry}{Entry clone}

\item{distination}{Distination vector}
}
\description{
TOPO reaction and LR reaction
}
\examples{
dstr_gw_topo("atgatgatgtga")

}
