% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-bstr.R
\name{print.bstr}
\alias{print.bstr}
\title{print biostrings class object}
\usage{
\method{print}{bstr}(x, ..., n = 6, l_name = 20, l_seq = 55)
}
\arguments{
\item{x}{x}

\item{...}{...}

\item{n}{number of printing elements}

\item{l_name}{max length of name}

\item{l_seq}{max length of sequence}
}
\description{
print biostrings class object
}
