% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{sort.bstr}
\alias{sort.bstr}
\title{sort bstr}
\usage{
\method{sort}{bstr}(x, decreasing = FALSE, ..., by = c("names",
  "length"))
}
\arguments{
\item{x}{x}

\item{decreasing}{logical. FALSE}

\item{...}{...}

\item{by}{c("names", "length")}
}
\description{
sort bstr
}
