% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{[.bstr}
\alias{[.bstr}
\title{subsetting bstr class object}
\usage{
\method{[}{bstr}(x, ..., drop = F)
}
\arguments{
\item{x}{x}

\item{...}{...}

\item{drop}{drop}
}
\description{
subsetting bstr class object
}
\examples{
(test <- dstr_rand_seq(3, 10, seed = 1))
test[1]
test[2:1]

test[2] <- "mutated"
test

}
