% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_io.R
\name{write_fasta}
\alias{write_fasta}
\title{write fasta file}
\usage{
write_fasta(x, fpath, width = 60)
}
\arguments{
\item{x}{x}

\item{fpath}{A fasta file path.}

\item{width}{w}
}
\description{
write fasta file
}
